plugin-sdk: split command status surface
