+++ src/plugin-sdk/command-status.ts
@@ -1,0 +1,5 @@
+export {
+  buildCommandsMessage,
+  buildCommandsMessagePaginated,
+  buildHelpMessage,
+} from "../auto-reply/status.js";
