+++ src/plugin-sdk/mattermost.ts
@@ -20,7 +20,7 @@ export {
   buildModelsProviderData,
   type ModelsProviderData,
 } from "../auto-reply/reply/commands-models.js";
-export { resolveStoredModelOverride } from "../auto-reply/reply/model-selection.js";
+export { resolveStoredModelOverride } from "../auto-reply/reply/stored-model-override.js";
 export {
   deleteAccountFromConfigSection,
   setAccountEnabledInConfigSection,
